moved classes concerning hierarchy and file naming to io.naming
