+++ src/test/groovy/edu/ie3/datamodel/io/naming/DefaultDirectoryHierarchyTest.groovy
@@ -3,9 +3,10 @@
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
  */
-package edu.ie3.datamodel.io.csv
+package edu.ie3.datamodel.io.naming
 
 import edu.ie3.datamodel.exceptions.FileException
+import edu.ie3.datamodel.io.naming.DefaultDirectoryHierarchy
 import edu.ie3.util.io.FileIOUtils
 import org.apache.commons.io.FilenameUtils
 import org.testcontainers.shaded.org.bouncycastle.util.test.TestFailedException
