+++ src/test/groovy/edu/ie3/datamodel/io/naming/FlatDirectoryHierarchyTest.groovy
@@ -3,8 +3,9 @@
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
  */
-package edu.ie3.datamodel.io.csv
+package edu.ie3.datamodel.io.naming
 
+import edu.ie3.datamodel.io.naming.FlatDirectoryHierarchy
 import edu.ie3.datamodel.models.input.system.BmInput
 import edu.ie3.util.io.FileIOUtils
 import spock.lang.Shared
