+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvThermalSourceTest.groovy
@@ -5,7 +5,7 @@
  */
 package edu.ie3.datamodel.io.source.csv
 
-import edu.ie3.datamodel.io.FileNamingStrategy
+import edu.ie3.datamodel.io.naming.FileNamingStrategy
 import edu.ie3.datamodel.io.factory.input.AssetInputEntityData
 import edu.ie3.datamodel.io.factory.input.ThermalUnitInputEntityData
 import edu.ie3.datamodel.models.input.OperatorInput
