+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvTypeSourceTest.groovy
@@ -5,7 +5,7 @@
  */
 package edu.ie3.datamodel.io.source.csv
 
-import edu.ie3.datamodel.io.FileNamingStrategy
+import edu.ie3.datamodel.io.naming.FileNamingStrategy
 import edu.ie3.datamodel.models.input.OperatorInput
 import spock.lang.Specification
 import edu.ie3.test.common.GridTestData as gtd
