+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvWeatherSourcePsdmTest.groovy
@@ -5,7 +5,7 @@
  */
 package edu.ie3.datamodel.io.source.csv
 
-import edu.ie3.datamodel.io.FileNamingStrategy
+import edu.ie3.datamodel.io.naming.FileNamingStrategy
 
 import static edu.ie3.datamodel.models.StandardUnits.SOLAR_IRRADIANCE
 import static edu.ie3.datamodel.models.StandardUnits.TEMPERATURE
